Refactoring utility files
