+++ app/src/main/java/swati4star/createpdf/util/ImageUtils.java
@@ -1,8 +1,16 @@
 package swati4star.createpdf.util;
 
+import android.graphics.Bitmap;
+import android.graphics.Canvas;
+import android.graphics.Color;
+import android.graphics.Paint;
+import android.graphics.PorterDuff;
+import android.graphics.PorterDuffXfermode;
+import android.graphics.Rect;
+
 import com.itextpdf.text.Rectangle;
 
-class ImageUtils {
+public class ImageUtils {
 
     /**
      * Calculates the optimum size for an image, such that it scales to fit whilst retaining its aspect ratio
@@ -26,6 +34,44 @@ class ImageUtils {
         float newHeight = originalHeight - (originalHeight * changeFactor);
 
         return new Rectangle(Math.abs((int) newWidth), Math.abs((int) newHeight));
+    }
+
+    /**
+     * Creates a rounded bitmap from any bitmap
+     * @param bmp - input bitmap
+     * @param radius - radius of output bitmap
+     * @return - output bitmap
+     */
+    public static Bitmap getRoundBitmap(Bitmap bmp, int radius) {
+        Bitmap sbmp;
+
+        if (bmp.getWidth() != radius || bmp.getHeight() != radius) {
+            float smallest = Math.min(bmp.getWidth(), bmp.getHeight());
+            float factor = smallest / radius;
+            sbmp = Bitmap.createScaledBitmap(bmp,
+                    (int) (bmp.getWidth() / factor),
+                    (int) (bmp.getHeight() / factor), false);
+        } else {
+            sbmp = bmp;
+        }
+
+        Bitmap output = Bitmap.createBitmap(radius, radius, Bitmap.Config.ARGB_8888);
+        Canvas canvas = new Canvas(output);
+
+        final Paint paint = new Paint();
+        final Rect rect = new Rect(0, 0, radius, radius);
+
+        paint.setAntiAlias(true);
+        paint.setFilterBitmap(true);
+        paint.setDither(true);
+        canvas.drawARGB(0, 0, 0, 0);
+        paint.setColor(Color.parseColor("#BAB399"));
+        canvas.drawCircle(radius / 2 + 0.7f, radius / 2 + 0.7f,
+                radius / 2 + 0.1f, paint);
+        paint.setXfermode(new PorterDuffXfermode(PorterDuff.Mode.SRC_IN));
+        canvas.drawBitmap(sbmp, rect, rect, paint);
+
+        return output;
     }
 
 }
