+++ app/src/main/java/swati4star/createpdf/util/MoveFilesToDirectory.java
@@ -30,12 +30,18 @@ public class MoveFilesToDirectory extends AsyncTask<String, String, String> {
     public static final int DELETE_DIRECTORY = 2;
     public static final int HOME_DIRECTORY = 3;
 
+    /**
+     * Inititlize async task to perform directory related operation
+     * @param context - context object
+     * @param filePath - list of file paths to be moved
+     * @param directoryName - directory name on which operation is to be performed
+     * @param mOperationID - operation id
+     */
     public MoveFilesToDirectory(Context context, ArrayList<String> filePath, String directoryName, int mOperationID ) {
         this.mContext = context;
         this.mFilePath = filePath;
         this.mDirectoryName = directoryName;
         this.mOperationID = mOperationID;
-
     }
 
     @Override
@@ -61,7 +67,6 @@ public class MoveFilesToDirectory extends AsyncTask<String, String, String> {
 
     @Override
     protected String doInBackground(String... strings) {
-
         switch (mOperationID) {
             case MOVE_FILES: {
                 String destination;
@@ -103,10 +108,16 @@ public class MoveFilesToDirectory extends AsyncTask<String, String, String> {
         return null;
     }
 
+    /**
+     * Moves files to a given directory
+     * @param directoryName - new directory
+     * @param source - source path
+     * @param destination - destination path
+     */
     private void moveFile( String directoryName , String source , String destination ) {
 
-        InputStream in = null;
-        OutputStream out = null;
+        InputStream in;
+        OutputStream out;
         int read;
         try {
             if (directoryName != null) {
