+++ app/src/main/java/swati4star/createpdf/util/PDFUtils.java
@@ -5,23 +5,13 @@ import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.net.Uri;
 import android.os.Environment;
-import android.support.design.widget.Snackbar;
-import android.text.Editable;
-import android.text.TextWatcher;
-import android.view.View;
-import android.widget.EditText;
 import android.widget.TextView;
 
-import com.afollestad.materialdialogs.DialogAction;
-import com.afollestad.materialdialogs.MaterialDialog;
 import com.itextpdf.text.Document;
 import com.itextpdf.text.DocumentException;
 import com.itextpdf.text.Element;
 import com.itextpdf.text.Font;
 import com.itextpdf.text.Paragraph;
-import com.itextpdf.text.exceptions.BadPasswordException;
-import com.itextpdf.text.pdf.PdfReader;
-import com.itextpdf.text.pdf.PdfStamper;
 import com.itextpdf.text.pdf.PdfWriter;
 
 import java.io.BufferedReader;
@@ -30,117 +20,31 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Objects;
 
 import swati4star.createpdf.R;
 import swati4star.createpdf.database.DatabaseHelper;
-import swati4star.createpdf.interfaces.DataSetChanged;
 import swati4star.createpdf.model.TextToPDFOptions;
 
 public class PDFUtils {
 
     private final Activity mContext;
-    private final FileUtils mFileUtils;
-    private String mPassword;
 
     public PDFUtils(Activity context) {
         this.mContext = context;
-        this.mFileUtils = new FileUtils(context);
-    }
-
-    /**
-     * Opens the password dialog to set Password for an existing PDF file.
-     *
-     * @param filePath Path of file to be encrypted
-     */
-    public void setPassword(final String filePath, final DataSetChanged dataSetChanged,
-                            final ArrayList<File> mFileList) {
-        final MaterialDialog dialog = new MaterialDialog.Builder(mContext)
-                .title(R.string.set_password)
-                .customView(R.layout.custom_dialog, true)
-                .positiveText(android.R.string.ok)
-                .negativeText(android.R.string.cancel)
-                .build();
-
-        final View mPositiveAction = dialog.getActionButton(DialogAction.POSITIVE);
-        assert dialog.getCustomView() != null;
-        EditText mPasswordInput = dialog.getCustomView().findViewById(R.id.password);
-        mPasswordInput.addTextChangedListener(
-                new TextWatcher() {
-                    @Override
-                    public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-                    }
-
-                    @Override
-                    public void onTextChanged(CharSequence s, int start, int before, int count) {
-                        mPositiveAction.setEnabled(s.toString().trim().length() > 0);
-                    }
-
-                    @Override
-                    public void afterTextChanged(Editable input) {
-                        if (StringUtils.isEmpty(input)) {
-                            showSnackbar(R.string.snackbar_password_cannot_be_blank);
-                        } else {
-                            mPassword = input.toString();
-
-                        }
-                    }
-                });
-        dialog.show();
-        mPositiveAction.setEnabled(false);
-        mPositiveAction.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                try {
-                    doEncryption(filePath, mPassword, mFileList);
-                    dataSetChanged.updateDataset();
-                    showSnackbar(R.string.password_added);
-                } catch (BadPasswordException e) {
-                    e.printStackTrace();
-                    showSnackbar(R.string.cannot_add_password);
-                } catch (IOException | DocumentException e) {
-                    e.printStackTrace();
-                }
-                dialog.dismiss();
-            }
-        });
-    }
-
-    /**
-     * Uses PDF Reader to set encryption in pdf file.
-     *
-     * @param path     - Path of pdf file to be encrypted
-     * @param password - password to be encrypted with
-     * @return string - path of output file
-     */
-    private String doEncryption(String path, String password,
-                                final ArrayList<File> mFileList) throws IOException, DocumentException {
-        String finalOutputFile = path.replace(".pdf", mContext.getString(R.string.encrypted_file));
-
-        if (mFileUtils.isFileExist(finalOutputFile)) {
-            int append = FileUtils.checkRepeat(finalOutputFile, mFileList);
-            finalOutputFile = finalOutputFile.replace(".pdf", append + ".pdf");
-        }
-
-        PdfReader reader = new PdfReader(path);
-        PdfStamper stamper = new PdfStamper(reader, new FileOutputStream(finalOutputFile));
-        stamper.setEncryption(password.getBytes(), mContext.getString(R.string.app_name).getBytes(),
-                PdfWriter.ALLOW_PRINTING | PdfWriter.ALLOW_COPY, PdfWriter.ENCRYPTION_AES_128);
-        stamper.close();
-        reader.close();
-        return finalOutputFile;
     }
 
     /**
      * Creates a dialog with details of given PDF file
      *
-     * @param name        - file name
-     * @param path        - file path
-     * @param size        - file size
-     * @param lastModDate - file's last modified date
+     * @param file        - file name
      */
-    public void showDetails(String name, String path, String size, String lastModDate) {
+    public void showDetails(File file) {
+
+        String name = file.getName();
+        String path = file.getPath();
+        String size = FileUtils.getFormattedSize(file);
+        String lastModDate = FileUtils.getFormattedSize(file);
+
         TextView message = new TextView(mContext);
         TextView title = new TextView(mContext);
         message.setText(String.format
@@ -165,97 +69,6 @@ public class PDFUtils {
         builder.show();
     }
 
-    /**
-     * Uses PDF Reader to decrypt the PDF.
-     *
-     * @param file Path of pdf file to be decrypted
-     */
-    public void removePassword(final String file,
-                               final DataSetChanged dataSetChanged,
-                               final ArrayList<File> mFileList) {
-        PdfReader reader = null;
-        try {
-            reader = new PdfReader(file, mContext.getString(R.string.app_name).getBytes());
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        //Check if PDF is encrypted or not.
-        if (!reader.isEncrypted()) {
-            showSnackbar(R.string.not_encrypted);
-            return;
-        }
-        final String[] input_password = new String[1];
-        final MaterialDialog dialog = new MaterialDialog.Builder(mContext)
-                .title(R.string.enter_password)
-                .customView(R.layout.custom_dialog, true)
-                .positiveText(android.R.string.ok)
-                .negativeText(android.R.string.cancel)
-                .build();
-        final View mPositiveAction = dialog.getActionButton(DialogAction.POSITIVE);
-        final EditText mPasswordInput = dialog.getCustomView().findViewById(R.id.password);
-        TextView text = dialog.getCustomView().findViewById(R.id.enter_password);
-        text.setText(R.string.decrypt_message);
-        mPasswordInput.addTextChangedListener(
-                new TextWatcher() {
-                    @Override
-                    public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-                    }
-
-                    @Override
-                    public void onTextChanged(CharSequence s, int start, int before, int count) {
-                        mPositiveAction.setEnabled(s.toString().trim().length() > 0);
-                    }
-
-                    @Override
-                    public void afterTextChanged(Editable input) {
-                        input_password[0] = input.toString();
-                    }
-                });
-        dialog.show();
-        mPositiveAction.setEnabled(false);
-        mPositiveAction.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                String finalOutputFile;
-                PdfReader reader = null;
-                try {
-                    reader = new PdfReader(file, mContext.getString(R.string.app_name).getBytes());
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-                byte[] password;
-                finalOutputFile = file.replace(".pdf", mContext.getString(R.string.decrypted_file));
-
-                if (mFileUtils.isFileExist(finalOutputFile)) {
-                    int append = FileUtils.checkRepeat(finalOutputFile, mFileList);
-                    finalOutputFile = finalOutputFile.replace(".pdf", append + ".pdf");
-                }
-
-                password = reader.computeUserPassword();
-                byte[] input = input_password[0].getBytes();
-                if (Arrays.equals(input, password)) {
-                    PdfStamper stamper = null;
-                    try {
-                        stamper = new PdfStamper(reader, new FileOutputStream(finalOutputFile));
-                    } catch (DocumentException | IOException e) {
-                        e.printStackTrace();
-                    }
-                    try {
-                        stamper.close();
-                    } catch (DocumentException | IOException e) {
-                        e.printStackTrace();
-                    }
-                    showSnackbar(R.string.password_remove);
-                    reader.close();
-                    dialog.dismiss();
-                    dataSetChanged.updateDataset();
-                } else {
-                    showSnackbar(R.string.incorrect_passowrd);
-                    dialog.dismiss();
-                }
-
-            }
-        });
-    }
 
     /**
      * Create a PDF from a Text File
@@ -311,12 +124,4 @@ public class PDFUtils {
         }
     }
 
-    /**
-     * Creates Snackbar
-     */
-    private void showSnackbar(int input) {
-        Snackbar.make(Objects.requireNonNull(mContext).findViewById(android.R.id.content),
-                input,
-                Snackbar.LENGTH_LONG).show();
-    }
 }
